removing crap and reorganizing code
