+++ src/env.mjs
@@ -9,6 +9,7 @@ export const env = createEnv({
   server: {
     // DATABASE_URL: z.string().url(),
     NODE_ENV: z.enum(["development", "test", "production"]),
+    GO_API_ENDPOINT: z.string().url(),
   },
 
   /**
@@ -27,6 +28,8 @@ export const env = createEnv({
   runtimeEnv: {
     // DATABASE_URL: process.env.DATABASE_URL,
     NODE_ENV: process.env.NODE_ENV,
+    GO_API_ENDPOINT: process.env.GO_API_ENDPOINT,
+
     // NEXT_PUBLIC_CLIENTVAR: process.env.NEXT_PUBLIC_CLIENTVAR,
   },
   /**
