--- src/pages/planets/Index.module.css
@@ -1,78 +1,0 @@
-.page {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    margin: auto;
-    max-width: 100vw;
-
-}
-
-.form {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: flex-start;
-    gap: 6px;
-    margin: auto;
-    width: 400px;
-    border: 1px solid gray;
-    border-radius: 12px;
-    padding: 10px;
-    margin-top: 30px;
-}
-
-.title {
-    width: 400px;
-    margin-top: 0;
-}
-
-.form > label {
-    font-size: 20px;
-    font-weight: 700;
-    margin-top: 8px;
-}
-
-.form > button {
-    margin-top: 10px;
-    background-color: rgb(41, 160, 41);
-    border: none;
-    color: #fff;
-    padding: 4px 10px;
-    cursor: pointer;
-}
-
-.table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-bottom: 20px;
-    border: 1px solid #ddd;
-  }
-  
-  .th,
-  .td {
-    padding: 8px;
-    text-align: left;
-    border-bottom: 1px solid #ddd;
-    width: 100%;
-    
-  }
-
-  
-  .th {
-    background-color: #f2f2f2;
-    font-weight: bold;
-  }
-  
-  /* CSS for table rows */
-  .tr:nth-child(even) {
-    background-color: #f9f9f9;
-  }
-  
-  .tr:hover {
-    background-color: #e5e5e5;
-  }
-  
-  /* CSS for table container */
-  /* CSS for table container */
