+++ src/server/api/root.ts
@@ -1,8 +1,4 @@
-import { exampleRouter } from "~/server/api/routers/example";
 import { createTRPCRouter } from "~/server/api/trpc";
-import { starsRouter } from "./routers/stars";
-import { planetsRouter } from "~/server/api/routers/planets";
-import { housesRouter } from "~/server/api/routers/houses";
 import { chartRouter } from "./routers/chart";
 
 /**
@@ -11,10 +7,6 @@ import { chartRouter } from "./routers/chart";
  * All routers added in /api/routers should be manually added here.
  */
 export const appRouter = createTRPCRouter({
-  example: exampleRouter,
-  stars: starsRouter,
-  planets: planetsRouter,
-  houses: housesRouter,
   chart: chartRouter
 });
