+++ src/utils/external/houses/types.d.ts
@@ -1,0 +1,13 @@
+export type house = {
+    name: string;
+    position: number;
+    sign: string;
+    longDegree: number;
+    longMinute: number;
+    longSecond: number;
+  };
+  
+  export type houseAPI = {
+    name: string;
+    longitude: string;
+  }
