+++ src/utils/external/stars/properties.ts
@@ -1,0 +1,56 @@
+export const majorStars = [
+    "Menkar",
+    "BatenKaitos",
+    "DenebKaitos",
+    "Difda",
+    "Scheat",
+    "Markab",
+    "Algenib",
+    "Enif",
+    "Hamal",
+    "Sheratan",
+    "Aldebaran",
+    "ElNath",
+    "AlHecka",
+    "Alcyone",
+    "Pollux",
+    "Castor",
+    "Sirius",
+    "Procyon",
+    "Praesepe",
+    "AsellusAustralis",
+    "AsellusBorealis",
+    "Betelgeuse",
+    "Bellatrix",
+    "Rigel",
+    "Saiph",
+    "Alpheratz",
+    "Almach",
+    "Mirach",
+    "Algol",
+    "Mirfak",
+    "Capella",
+    "Menkalinan",
+    "Regulus",
+    "Denebola",
+    "Zosma",
+    "Spica",
+    "Acrux",
+    "Vindemiatrix",
+    "Algorab",
+    "Unukalhai",
+    "Toliman",
+    "Agena",
+    "Zubenelgenubi",
+    "Zubeneshamali",
+    "Antares",
+    "Shaula",
+    "Aculeus",
+    "Acumen",
+    "Alphard",
+    "Rasalgethi",
+    "Ascella",
+    "Facies",
+    "Nunki",
+    "Vega",
+];
