+++ src/utils/external/stars/types.d.ts
@@ -1,0 +1,25 @@
+export type majorStarResponse = {
+    starName: string;
+    longitude: string;
+    latitude: string;
+    house: string;
+    distance: string;
+    speed: string;
+    magnitude: string;
+    altName: string;
+}
+
+export type majorStar = {
+    star: string;
+    constellation: string;
+    position: number;
+    sign: string;
+    longDegree: number;
+    longMinute: number;
+    longSecond: number;
+    lat: number;
+    speed: number;
+    house: number;
+    distance: number;
+    magnitude: number;
+  }
